requirements.txt
absl-py==1.4.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
filelock==3.12.4
google-auth==2.23.0
google-auth-oauthlib==1.0.0
grpcio==1.58.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
Markdown==3.4.4
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
oauthlib==3.2.2
Pillow==10.0.1
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.7
zipp==3.16.2
